split functions into right files
